backend/requirements.txt
# FastAPI and web
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
pgvector==0.2.5

# Redis
redis==5.0.1

# Integrations
plaid-python>=14.0.0
stripe>=7.0.0
python-multipart>=0.0.9

# AI / LLM orchestration
anthropic>=0.40.0
openai>=1.30.0
google-genai>=1.0.0
langgraph>=0.2.0
langchain-core>=0.3.0

# Finance data
httpx>=0.26.0          # async HTTP for Finance MCP (Finnhub, AV, NewsAPI)

# Retrieval
rank-bm25>=0.2.2       # BM25-Okapi — KnowledgeScanner (optional, see knowledge_scanner.py)

# Utilities
python-dotenv==1.0.1
httpx==0.26.0
tenacity==8.2.3
